fix(dashboard): wire ServerGreeting into dashboard page
ServerGreeting was a fully-built async server component that was never
rendered anywhere in the application. This fix wires it into the
dashboard page as a personalized welcome banner.

Changes:
- Convert page.tsx from a client component to a server component
- Extract client-side dashboard logic into DashboardClient.tsx
- Import and render ServerGreeting above the dashboard client content
- Update a11y test to work with the refactored component structure

Closes #976
